Remove metabase/lib/schema dependency
